# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "Документация API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/saved-wishlists": {
            "get": {
                "tags": [
                    "Saved wishlists"
                ],
                "summary": "Чужие списки пользователя, последние добавленные — первыми",
                "operationId": "1b5df88744eb1f5b5e3905bdb5c678bb",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SavedWishlist"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/saved-wishlists/{id}": {
            "post": {
                "tags": [
                    "Saved wishlists"
                ],
                "summary": "Добавить чужой список к себе",
                "description": "Добавить можно список желаний или список дел с доступом по ссылке. Повторное добавление возвращает прежнюю запись",
                "operationId": "81e18fdd578ebee1a3ea49d6f9b3a047",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID списка",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список добавлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SavedWishlist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Список не найден или не открывается по ссылке (заметка, закрытый список дел)"
                    },
                    "422": {
                        "description": "Свой список нельзя добавить в чужие"
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Saved wishlists"
                ],
                "summary": "Убрать список из чужих",
                "description": "Отсутствующая запись не считается ошибкой",
                "operationId": "8553567bd7bc2be8d7b0df0436e9355d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID списка",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список убран",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "01jq3v7x8k2m4n6p8r0t2w4y6z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SavedWishlist": {
                "description": "Сводка чужого списка. Если владелец закрыл доступ к списку дел, available = false и поля списка равны null",
                "properties": {
                    "id": {
                        "description": "ID самого списка",
                        "type": "string",
                        "example": "01jq3v7x8k2m4n6p8r0t2w4y6z"
                    },
                    "available": {
                        "description": "Открывается ли список по ссылке",
                        "type": "boolean",
                        "example": true
                    },
                    "savedAt": {
                        "description": "Когда список добавлен",
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-10-10T09:15:00+00:00"
                    },
                    "username": {
                        "description": "Имя владельца",
                        "type": "string",
                        "example": "anna",
                        "nullable": true
                    },
                    "avatarUrl": {
                        "description": "Фотография владельца; null — показывается первая буква имени",
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "example": "gift",
                        "nullable": true,
                        "enum": [
                            "gift",
                            "todo"
                        ]
                    },
                    "title": {
                        "type": "string",
                        "example": "День рождения",
                        "nullable": true
                    },
                    "color": {
                        "type": "string",
                        "example": "mint",
                        "nullable": true,
                        "enum": [
                            "white",
                            "lavender",
                            "pink",
                            "peach",
                            "lemon",
                            "mint",
                            "sky"
                        ]
                    },
                    "dueDate": {
                        "description": "Срок списка дел или дата события списка желаний (Y-m-d)",
                        "type": "string",
                        "format": "date",
                        "example": "2026-12-31",
                        "nullable": true
                    },
                    "itemsCount": {
                        "description": "Число позиций",
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "selectedCount": {
                        "description": "Выбранные гостями подарки или выполненные дела",
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}